Minor improvements

+++ d3/turtle.cpp
@@ -47,7 +47,7 @@ void Turtle::left() {
 }
 
 void Turtle::write(Vectorfield vf) {
-    vf.set(_x, _y, _direction);
+    vf.set((size_t)_x, (size_t)_y, _direction);
 }
 
 void Turtle::move_turn_write(Vectorfield vf, bool turn_left_after_moving) {
